+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Added the ability to plot multiple fitting histories returned by the [`fit`](https://keras.io/api/models/model_training_apis/#fit-method) function on a single figure with the `lcurves_by_history` function.
 - Added such parameters for the `lcurves_by_history` function:
-  - `color_grouping_by` which specifies how colors of curves in the subplots are grouped;
+  - `color_grouping_by` which specifies how colors of curves in subplots are grouped;
   - `model_names` which specifies model names for each history in the `history` list to use in the legends of the subplots;
   - `optimization_modes` which specifies optimization modes of metrics ("min" or "max").
 - Added the `utils` module with two functions: `get_mode_by_metric_name` and `get_best_epoch_value`.